requirements.txt
asgiref==3.7.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
Django==4.2.6
google-auth==2.23.3
idna==3.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
requests==2.31.0
rsa==4.9
sqlparse==0.4.4
urllib3==2.0.6
